internal: machinery for task run identification (#16)
# The problem

A couple of `dorq` commands require identifying the task run. There's a
couple of ways we can figure it out based on user input. The same code
is likely to be used by different commands.

# This PR's solution

Implements Argument Resolver for finding task run ID.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box). I'm not adding a changelog entry
because this PR's content will be only visible to the user when we
implement task-related commands in the future.
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`