fix!: inconsistent data shapes when getting intermediate artifacts (#23)
# The problem

* The contents of the dict returned from `wf_run.get_results()` depended
on the runtime type used. For Ray and QE, this was a map if [task inv
id]:[artifact value]. For In-Process runtime this was [artifact
ID]:[artifact value].
* `InProcessRuntime` didn't conform to `RuntimeInterface`. The types and
semantics didn't always match.
* The docstrings for `sdk.WorkflowRun.get_results()` mentioned
`TaskRunID`s but in practice this was never used. Instead, we would use
`TaskInvocationID` or `ArtifactID` as the filter and in the returned
dict.

# This PR's solution

* Marks `InProcessRuntime` as an implementation of `RuntimeInterface`.
* Fixes invalid data shape returned from
`InProcessRuntime.get_available_outputs()`.
* Fixes docstrings for `sdk.WorkflowRun.get_results()`.
* Simplifies `sdk.WorkflowRun.get_results()`: now it doesn't accept any
arguments. Turns out it would be increasingly difficult to test after
this PR + the client-side filtering seems to have limited user value.
* Fixes type annotations & updates tests for
`CERuntime.get_available_outputs()`

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`