feat: dorq wf list
# The problem

Users need to be able to see a filtered list of workflow runs via the
cli.

# This PR's solution

- user facing:
  - Implements `dorq wf list` with the following arguments
- `-c` / `--config` specifies a config for which workflows will be
listed. Can be specified multiple times. If omitted, the user is
prompted to select from a checklist of configs.
- `-a` / `--all` show all workflow runs that match the specified
filters. incompatible with `-i`.
- `-i` / `--interactive` prompt the user for any filters that are not
specified. Incompatible with `-a`.
    - filters:
- filters are applied per config, so if the user specifies 2 configs and
a limit of 10, up to 20 workflow runs will be displayed.
- `-l` / `--limit` set the maximum number of runs to display per config.
      - `-t` / `--max-age` set the maximum age of runs to display.
- `-s` / `--state` only show workflow runs of the specified state. Can
be specified multiple times. If one or more specified states are not
valid workflow run states, prompts the user to choose from available
states.
- non-user facing:
- extends prompter to allow prompting for checkboxes, and integer
inputs.
- refactors workflowrunrepo to make use of the API `list_workflow_runs`
method.
- extends workflowrunrepo to provide `list_wf_runs` method. The
preexisting `list_wf_run_ids` method now uses this.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`

[ORQSDK-671]


[ORQSDK-671]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-671?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: James Clark <70290797+jamesclark-Zapata@users.noreply.github.com>