fix!: remove fluentbit (#63)
# The problem

We've had two ways of reading local runtime logs: directly and using
FluentBit. The FluentBit integration was never fully implemented and
introduced dependence on docker. The main reason for using FluentBit was
log reading performance, but this improvement was never confirmed in
practice.

On top of that, the recent series of logs-related bugfixes were only
applied to the direct log reading. Fixing the FluentBit integration
doesn't look like a good investment of our time. Also, having a flaky
functionality is a release blocker.

# This PR's solution

* Removes code for managing FluentBit via docker compose.
* Removes code for reading files produced by FluentBit.
* Removes CLI options for starting and stopping the FluentBit container:
`orq {up,down} --fluentbit`.
* Removes docs for setting up the FluentBit pre-requirements.
* Removes the old `corq services ...` CLI code. It's not used by Studio,
and it's not exposed in the public API, so it shouldn't be used by our
users.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`