imp: print when polling when awaiting wf completion (#69)
# The problem

I've been writing some scripts
In every script for submitting and awaiting workflows I need to await
the workflow completion. We have `sdk.WorkflowRun.wait_until_finished()`
but when the workflow is big, or it's running on the remote runtime, it
can take long with no information for the user.

# This PR's solution

I've added a `verbose=True` flag to `.wait_until_finished()`. It prints
to stderr in every iteration of the polling loop.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`