feat: raise an exception if workflows have zero tasks. (#86)
[ORQSDK-774]

# The problem

On QE you get an HTTP error stating that you need at least 1 task to be
executed if your WF doesn't use any tasks. This is not the case on local
Ray or CE. We should ensure that this is consistent across all runtimes.

# This PR's solution

- Adds a check during traversal that raises a WorkflowSyntaxError when
no task invocations are identified.
- The exception reads: "The workflow '[wf name]' cannot be submitted as
it does not define any tasks to be executed. Please modify the workflow
definition to define at least one task and retry submitting the
workflow."
- If the source file and line can be identified, these are also included
in the message.

# Checklist

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`


[ORQSDK-774]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-774?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ