chore: limit github action workflows (#90)
# The problem

We've had some issues with Windows tests hanging for 6 hours (this might
be a `grpcio` issue, but that's something else to look at).

# This PR's solution

Adds:
- 15 minute limit to workflow jobs
- Concurrency limits based on the workflow and the ref.

`github.ref` docs:
> The fully-formed ref of the branch or tag that triggered the workflow
run. For workflows triggered by push, this is the branch or tag ref that
was pushed. For workflows triggered by pull_request, this is the pull
request merge branch. For workflows triggered by release, this is the
release tag created. For other triggers, this is the branch or tag ref
that triggered the workflow run. This is only set if a branch or tag is
available for the event type. The ref given is fully-formed, meaning
that for branches the format is refs/heads/<branch_name>, for pull
requests it is refs/pull/<pr_number>/merge, and for tags it is
refs/tags/<tag_name>. For example, refs/heads/feature-branch-1.

`github.workflow` docs:
> The name of the workflow. If the workflow file doesn't specify a name,
the value of this property is the full path of the workflow file in the
repository.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`