fix!: Runtimes return serialised values (#95)
Depends on #99

This is not ready for merge to main. @alexjuda and I were considering a
strategy to mitigate the breaking changes.

In the new CE runtime, when running on multiple nodes, workers may not
have the same Python environment installed. This caused workflow
failures.

* Restructures internal runtimes to return serialised values only
* Updates Ray runtime to handle serde inside task code
* Deserialises workflow/task results inside the API layer
* Adds a new JSON encoder to accurately serialise tuples

Misc:
* Bumps `typing_extensions` to `4.1.0` to get `assert_never` to annotate
unreachable code.

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`

---------

Co-authored-by: Alexander Juda <6004040+alexjuda@users.noreply.github.com>