refactor: `orquestra.sdk.packaging` module -> package (#108)
# The problem

When working on
https://github.com/zapatacomputing/orquestra-workflow-sdk/pull/105 I
needed to add some helpers related to fetching version of the SDK.
`orquestra.sdk.packaging` seemed like a good place for it. I don't want
to expose my helpers in the public API, but `orquestra.sdk.packaging` is
a module.

# This PR's solution

~Depends on
https://github.com/zapatacomputing/orquestra-workflow-sdk/pull/107.~

Turns `orquestra.sdk.packaging` into a package. This allows separating
public-internal APIs.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`