internal: Placeholder GitHub Actions workflow to publish orquestra-sdk-base Docker image (#138)
We need this on the main branch to be able to test the workflow on the
actual branch during development. The final version of the workflow will
probably look different.

Jira ticket: https://zapatacomputing.atlassian.net/browse/ORQP-1476

# The problem

The GitHub actions workflow for publishing orquestra-base-sdk Docker
images is not visible in the GitHub UI.

# This PR's solution

Add a placeholder version of the GitHub actions workflow to publish
Docker images into the main branch until development of it is completed
and it's replaced with the proper working version.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).