fix: Set max retries in Ray (#140)
# The problem

When a Ray worker crashes, it will try to run the task again. This
causes issues when using things that assume a single execution of a
task.

# This PR's solution

* Set `max_retries` to 0.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).