internal: Tagging Docker images via optional parameter [OP-39] (#145)
# The problem

It's not possible to tag Docker images based on the SDK version with the
current version of the GitHub Actions workflow that we have for
publishing Docker images.

# This PR's solution

Add an optional parameter to the GitHub Actions workflow that publishes
Docker images to optionally tag the published Docker image. This way,
when we want to publish a Docker image and also tag it with the version
of SDK that it uses, we'll be able to do so.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).