Feat: Prompt for workspace and project in all CLI commands that prompted for WF id (#149)
# The problem

With workspaces and projects introduction, all workflows are beloning to
given workspace and project. User need to have a way to chose from which
workspaces and projects does they want to list workflows

# This PR's solution

after prompting for config, prompt for workspace and project

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [ ] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).