feat: login with config name as alternative to URI
As an alternative to
```
orq wf login -s https://whatever.orquestra.blah
```
This allows for
```
orq wf login -c bonfig_name
```

The action then loads the specified config and uses the URI stored therein.

If the `--ce` flag is not used, the login is performed for the same runtime as the stored config. If the `--ce` flag _is_ used, we check the config for a mismatch and raise an error if they differ.