internal: Fix CUDA image publishing workflow failure when no Docker tag is given (#152)
# The problem

When no Docker tag is given as an input parameter, the GitHub Actions
workflow that builds CUDA images fails. However, it needs to be an
optional input parameter and the image should be tagged with prefix
`dev-build` instead.

# This PR's solution

Modify GitHub Actions workflow code to add the additional tags only when
they're provided.

# Checklist

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).