Feat: cli easy re login (#154)
# The problem

If I want to log in to a config I've already used, I typically have a)
have memorized the uri, or b) open the config.json to check the uri.
It's also very easy to forget the `--ce` flag and have to log in _again_
once I notice.

# This PR's solution

Adds a config option to orq login, so users can now do:

```
orq login -c my_config
```

the uri and CE/not CE are then read from the saved config, and login
proceeds with those parameters.

If the token is already known, then updating it is equally simple:

```
orq login -c my_config -t my_token
```

This creates a number of potential clashes which are mitigated as
follows:

| case | mitigation |
|------|-----------|
| uri and config both specified | The cli requires exactly one argument
out of `-c` and `-s`, and raises an exception otherwise.|
| `--ce` with a non-CE stored config | Raise an exception informing the
user fo the clash and giving the precise command to change the runtime
in the stored config (`orq login` with the uri sourced from the stored
config) if that's what's intended. |
| `--ce` not flagged with a CE config | In the absence of explicit
action from the user, we assume the stored config is authoritative, so
proceed as though `--ce` had been flagged |

# Checklist

_Check that this PR satisfies the following items:_

- [X] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [X] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [X] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).