feat: Check tokens before saving in orq (#158)
# The problem

Sometime folks were pasting in the wrong thing as a token and weren't
notified until after they attempt to use the API

# This PR's solution

Adds a *non-cryptographic* check of the token.

After checking:
* the token is a JWT
* it is not expired

What this explicitly does not check:
* RSA signature
* Claims

This is OK because the server does the final auth check, we're just
doing a sanity check that the JWT is of the right shape. If folks still
have issues, we can fetch the public key and do the signature check

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).