feat: give more informative error message when ray cluster isn't running (#163)
# The problem

When user tries to use local “ray” config, and ray isnt running, the
error message that pops up is coming from ray itself saying something
along “ray cluster not running bla bla bla”

We should catch that error and translate it to more informative message
- to see if ray is running and if not, to start it with orq up command

# This PR's solution

Adds `RayNotStartedError` which is raised from the ConnectionError when
the client tries to connect to ray and cannot find a running instance.

example:
```
> orq wf submit scratch.py -c ray
Could not find any running Ray instance. You can use 'orq status' to check the status of the ray service. If it is not running, it can be started with the `orq up` command.
```

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).

[ORQSDK-828](https://zapatacomputing.atlassian.net/browse/ORQSDK-828)

[ORQSDK-828]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-828?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: James Clark <70290797+jamesclark-Zapata@users.noreply.github.com>