internal: Add a makefile for building the sdk images, tweak for building on M1 (#159)
# The problem

Studio development requires building images locally, which is _much_
faster if it's an ARM-native image. Also, Makefiles remove the need to
remember `docker build` commands.

Thirdly, workflow dispatches have a limit of 100 chars on the
`event_type` field 🥲

# This PR's solution

Added a Makefile, and installed a few more dependencies in the image
build, conditional on the `TARGETARCH` [automatic
`ARG`](https://docs.docker.com/engine/reference/builder/#automatic-platform-args-in-the-global-scope).
Tweaked the `event_type` field in the CICD-actions dispatches.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).