Imp: Improve workspace and project prompters in studio showing and auto-selecting current context (#182)
# The problem

When in studio, SDK is aware of workspace and project. This info can be
used to help users interact with current context
https://zapatacomputing.atlassian.net/browse/ORQSDK-808

# This PR's solution

When in studio - when SDK prompts for workspace and project, Current
context is auto-selected + highlighted with additional info

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/internal/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).