docs: Update documentation as we disable single node runtime (#197)
# The problem

We have disabled single node runtime and by default provision a Ray
cluster, even when nodes=1 for the workflow resources. The documentation
still mentions that the nodes parameter needing to be set to a value
larger than 1, which is no longer the case.

# This PR's solution

Update the documentation to explain the current situation correctly.

# Checklist

- [X] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [X] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [X] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).