Fix: Fix Ray WFs failing caused by any task returning dict defined in return statement (#205)
# The problem

When the task had dict defined in return statement it caused the whole
workflow to fail. Caused by the fact that we treated dicts as
subscribtable (with [] operator) which they are not.

# This PR's solution

Do not treat dict return values as subscribtable

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).