Feat!: Changed wf_def.get_tasks() return type from set, to topologically sorted list (#222)
# The problem

get_tasks() was returning `set()` object which was confusing for a lot
of users
https://zapatacomputing.atlassian.net/browse/ORQSDK-885

# This PR's solution

Change returned type to list, which is sorted topologically. 

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).