docs: correct references to 'in-process' that should be 'in_process' (#239)
# The problem

The help text for the config option in our CLI directs the user to use
`in-process` for the in-process runtime. The actual argument they should
use is `in_process`.

# This PR's solution

Corrects this help string, and the only other reference I could find to
this name.

# Checklist

_Check that this PR satisfies the following items:_

- [X] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [X] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [X] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [X] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).