docs: rm custom-images guide (#271)
# The problem

As the custom image on CE process will be handled by platform team, it
was decided by The Powers That Be (@zaptim and @roryschadler) to move
this to internal documentation.

# This PR's solution

Remove the `.rst` files from the source. The content has already been
transferred to internal docs.

# Checklist

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).