feat: remove checks for additional env vars when determining local or remote execution. (#270)
[ORQSDK-914]

# The problem

We check for a bunch of environment variables when determining whether
execution is local or remote. This was to provide backwards
compatibility to platform versions that lacked the CURRENT_CLUSTER_ENV
variable. Since we only need to support N-1 versions, we can safely rely
on CURRENT_CLUSTER_ENV going forward.

# This PR's solution

Removes the redundant checks.

# Checklist

_Check that this PR satisfies the following items:_

- [X] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [X] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [X] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [X] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).


[ORQSDK-914]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-914?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ