int: prototype makefile cleanup (#283)
# The problem

Our nested makefile with z-quantum-actions subtree is outdated and needs
an overhaul. The original idea [described on
Jira](https://zapatacomputing.atlassian.net/browse/ORQSDK-456) was to
replace the subtree with a makefile fetched from another source.

Meanwhile, it doesn't make a lot of sense either. The makefile in this
repo overrides all targets anyway (the only exception: `make clean`).

# This PR's solution

POC of an alternative solution, other than the one originally specified
in the ticket.

* Removes the original makefile inheritance.
* Adds missing make variables.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).