docs: simple docs fixes (#292)
# The problem

We have several tickets relating to simple wording or formatting changes
in the docs.

# This PR's solution

* General
* cffd5ec244b238686cfc4b1b0d1a42577fd459ef &
[235e98b](https://github.com/zapatacomputing/orquestra-workflow-sdk/pull/292/commits/235e98bc4be49960b30fcc7fd29054bd9b711547)
* Reformat docs so that each sentence is on its own line to keep future
diffs short.
* [ORQSDK-894]
*
[5f895a3](https://github.com/zapatacomputing/orquestra-workflow-sdk/pull/292/commits/5f895a3555d7ec554010da0fa05ced1dd2395099)
* Only specify Python3.9 in installation instructions as that's what CE
requires.
* [ORQSDK-897]
*
[2fa350f](https://github.com/zapatacomputing/orquestra-workflow-sdk/pull/292/commits/2fa350fb380f337f51efa1fc3d6d859d0a66750a)
* “Examples given in this tutorial will run those workflows in a local
Ray instance.” → “Example workflows given in this tutorial will be run
in a local Ray instance”
* "As an example of parametrized workflow with 1 task, you can define
one as follows:” → “Here is an example of a parametrized workflow with
one task:”
* “To get the calculated values, either follow the instruction in
[started
Ray](https://docs.orquestra.io/docs/core/sdk/tutorials/ray.html). or run
the following” → “To get the calculated values, run the following”
* “The biggest advantage of parametrized workflow is to run multiple
versions of the same workflow to compare results for different inputs.
This can be obtained by executing workflow with different input
parameters in a loop as in this example:” → “The biggest advantage of
parametrized workflows is the ability to run multiple versions of the
same workflow to compare results for different inputs. This can be
performed by executing the workflow with different input parameters in a
loop as in this example:”
* [ORQSDK-900] 
    * 5da4c1b138fbc0a601cc3f5561673cb1add0bd63 
* “the same process, the creates WorkflowRun” → “the same process, the
latter creates WorkflowRun”
      *  Fixed incorrect rendering of builtin configurations list
      * ”confiuration” → “configuration”
* “Configurations control interaction with runtime backend” →
“Configurations control interaction with the runtime backend”
* “Local ray runtime has hardcoded config name” → “The local ray runtime
has a hardcoded config name”
* “In-process runtime has also hardcoded config name” → “The in-process
runtime also has a hardcoded config name”
* [ORQSDK-901]
  * a299ff2a6c84c4294eff37343a8f8ddad76757c3
* “This is what you should use in the CLI to reference a configuration”
→ “This is what you should use in the CLI or SDK to reference a
configuration”
    * added link to CLI reference so users can click there.
  * bcb729d54c1a030255a4b58a01ca5ac8b7748986
* "Currently there are three supported runtimes: local execution and
remote execution via Compute Engine." -> "Currently there are two
supported runtimes: local execution via Ray, and remote execution via
Compute Engine."
* [ORQSDK-902]
  * 4f7ecea4910ee4243f3b0fb8accacf8164365513
* Rework the paragraph that introduces the gotchas list so that it reads
less like we don't care. The new text is: "The following are examples of
common gotchas that our users have run into, and the workarounds and/or
prescribed methods of achieving the desired behaviour. Note that many of
these cases will be the subject of future work, so the behaviour
documented here is liable to change."
*
[ORQSDK-904](https://zapatacomputing.atlassian.net/browse/ORQSDK-904?atlOrigin=eyJpIjoiN2YzM2ZiMTZjYmEyNGVmNjllZWRkMDVmMmJlNjYxZTYiLCJwIjoiaiJ9)
  * df999b42db67d8132aca4dbb5c798476b88cf915
* “When you request more than one nodes” → “When you request more than
one node”
* “Resources that you request for your workflow needs to be larger than
what you request” → “The resources that you request for your workflow
need to be larger than what you request”

# Checklist

_Check that this PR satisfies the following items:_

- [X] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [X] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [X] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).


[ORQSDK-900]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-900?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ORQSDK-901]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-901?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ORQSDK-902]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-902?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[ORQSDK-897]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-897?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[ORQSDK-894]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-894?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: James Clark <70290797+jamesclark-Zapata@users.noreply.github.com>